extract Cache and AsyncCache interfaces into their own file
